WIP: Crear archivos run_x
