+++ .gitignore
@@ -7,6 +7,7 @@ Modelos*
 resultado.mp4
 Resultados_entrenamiento*
 Resultados_reportes*
+results*
