+++ config.yaml
@@ -12,16 +12,16 @@ deteccion:
   un_video:
     procesar: False #! Procesar un video en específico
     guardar: True #! Guardar el video con los resultados
+    zona: "Zona C" #! Zona a considerar
+    path_destino: "results/deteccion/" #! Carpeta donde se guardará el video con los resultados
     # path_origen: "dataset/dataset_original/Zona A/20240102_131846.mp4"    #! Path del video a procesar
     path_origen: "dataset/dataset_original/Zona C/20240102_132223.mp4" #! Path del video a procesar
     # path_origen: "dataset/dataset_reescalado-576x1024-5fps/Zona A/20240102_131846.mp4"    #! Path del video a procesar
-    zona: "Zona C" #! Zona a considerar
-    path_destino: "results/deteccion/" #! Carpeta donde se guardará el video con los resultados
 
   procesar_camara: False #! Procesar la cámara en tiempo real
 
 decision:
-  decision: False #* Iniciar la toma de decisiones
+  decision: True #* Iniciar la toma de decisiones
   ponderaciones_zonas: [
       1.0,
       1.0,
@@ -40,8 +40,9 @@ decision:
 
   # path_modelo_entrenado: "results/training/DQN_2024-04-21_19-55/epoca_18.h5"    #! No funciona: 1235
   # path_modelo_entrenado: "results/training-2/DQN_2024-05-02_15-33/epoca_14.h5"  #! Si funciona: 1322
-  path_modelo_entrenado: "results/training/DQN_2024-05-06_12-21/epoca_13.h5" #! Si funciona: 786
+  # path_modelo_entrenado: "results/training/DQN_2024-05-06_12-21/epoca_13.h5" #! Si funciona: 786 ES EL BUENO
   # path_modelo_entrenado: "results/training/DQN_2024-05-16_11-06/epoca_25.h5"    #! Si funciona: 895
+  path_modelo_entrenado: "assets/dqn_models/2024-05-06_12-21-epoca_13.h5"
 
   entrenamiento:
     entrenar: False #* Entrenar el modelo
